[Chore] #66 - book과 관련된 패키지 이동
