--- src/main/java/com/arabook/arabook/api/book/service/BookService.java
@@ -1,10 +1,0 @@
-package com.arabook.arabook.api.book.service;
-
-import com.arabook.arabook.api.book.controller.dto.response.BookDetailResponse;
-import com.arabook.arabook.api.book.controller.dto.response.BooksResponse;
-
-public interface BookService {
-  BooksResponse getBooksBySearch(String keyword);
-
-  BookDetailResponse getBookDetail(Long bookId);
-}
