+++ src/main/java/com/arabook/arabook/book/controller/RecommendController.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.api.book.controller;
+package com.arabook.arabook.book.controller;
 
 import static com.arabook.arabook.common.success.book.BookSuccessType.*;
 
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.arabook.arabook.api.book.controller.dto.response.AIRecommendBookResponse;
-import com.arabook.arabook.api.book.controller.dto.response.BooksResponse;
-import com.arabook.arabook.api.book.service.BestSellerService;
 import com.arabook.arabook.api.global.security.AuthMember;
+import com.arabook.arabook.book.controller.dto.response.AIRecommendBookResponse;
+import com.arabook.arabook.book.controller.dto.response.BooksResponse;
+import com.arabook.arabook.book.service.BestSellerService;
 import com.arabook.arabook.common.response.ResponseTemplate;
 
 import lombok.RequiredArgsConstructor;
