+++ src/main/java/com/arabook/arabook/book/controller/dto/response/BookDetailResponse.java
@@ -1,11 +1,11 @@
-package com.arabook.arabook.api.book.controller.dto.response;
+package com.arabook.arabook.book.controller.dto.response;
 
 import java.time.Year;
 import java.util.List;
 
 import com.arabook.arabook.api.category.controller.dto.response.SubCategoryResponse;
 import com.arabook.arabook.api.hashtag.controller.dto.response.HashTagResponse;
-import com.arabook.arabook.storage.domain.book.entity.Book;
+import com.arabook.arabook.book.entity.Book;
 
 import io.swagger.v3.oas.annotations.media.Schema;
