+++ src/main/java/com/arabook/arabook/book/entity/BookHashtagMapping.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.storage.domain.book.entity;
+package com.arabook.arabook.book.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -9,7 +9,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
-import com.arabook.arabook.storage.domain.category.entity.SubCategory;
+import com.arabook.arabook.storage.domain.hashtag.entity.Hashtag;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -18,17 +18,17 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "book_sub_category_mappings")
-public class BookSubCategoryMapping {
+@Table(name = "book_hashtag_mappings")
+public class BookHashtagMapping {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long bookSubCategoryMappingId;
+  private Long bookHashtagMappingId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "book_id")
   private Book book;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "sub_category_id")
-  private SubCategory subCategory;
+  @JoinColumn(name = "hashtag_id")
+  private Hashtag hashtag;
 }
