+++ src/main/java/com/arabook/arabook/book/repository/BookRepository.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.storage.domain.book.repository;
+package com.arabook.arabook.book.repository;
 
 import static com.arabook.arabook.common.exception.book.BookExceptionType.*;
 
@@ -6,8 +6,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.arabook.arabook.book.entity.Book;
 import com.arabook.arabook.common.exception.book.BookException;
-import com.arabook.arabook.storage.domain.book.entity.Book;
 
 public interface BookRepository extends JpaRepository<Book, Long>, BookCustomRepository {
   Optional<Book> findByBookId(Long bookId);
