+++ src/main/java/com/arabook/arabook/book/service/BestSellerService.java
@@ -1,0 +1,10 @@
+package com.arabook.arabook.book.service;
+
+import com.arabook.arabook.book.controller.dto.response.AIRecommendBookResponse;
+import com.arabook.arabook.book.controller.dto.response.BooksResponse;
+
+public interface BestSellerService {
+  BooksResponse getBestSellerBooks();
+
+  AIRecommendBookResponse getAIRecommendBook(Long memberId);
+}
